Add a top-level `include` dir for public headers, AT-SPI for now
This is fairly standarized pattern for libraries: headers meant to be
installed and used publicly by the library users go in an `include`
dir, while all other (private) headers go in their respectives dirs
alongside the sources that use them.

My understanding is that we will have the following structure for
public headers, to be installed in `${PREFIX}/include`

- axaccess/ where the main, platform-agnostic headers go
  - atspi/    AT-SPI specific headers
  - (a directory for the other backends, one for each)

Then a program would do, for example:

```
 #include <axaccess/AXAaccess.h>
 #include <axaccess/atspi/atspi_node.h>
 [...]
```

The reason to have `axaccess` prefix is to avoid collision with
potentially similarly named headers from other libraries. That's a good
practice in general.