[ATSPI] Add feature flag for python bindings too
Similar to what we did for NodeJS, we put the python3 bindings under a
flag that is ON by default, so that users that don't need this binding
are not annoyed by the dependency.

Now swig is required and imported only if we have either Python or NodeJS
bindings ON.

The README file was updated accordingly.