Turn axa::NodePtr into a unique pointer instead of shared
The main reason for this is to be able to build NodeJS bindings via
SWIG. Right now that doesn't work for shared pointers.

This move is risky and could bring issues down the line if we ever need
to keep more than one reference to a node. In that case, there are some
possible solutions, like copying the object or creating a new one and
increasing the reference count of internal members.