[docs] Add initial autogenerated Doxygen documentation (#120)
* [docs] Add initial autogenerated Doxygen documentation

This is mostly CMake Doxygen infrastructure, with the minimum code
documentation to devide the different APIs in groups, so that they
appear in separate modules in the documentation.

The idea is that follow-up patches from API "owners" will complete the
code annotations, so that each API is documented in a way that is most
convenient/coherent for the API in question.

There is also some general, introductory documentation missing in the
main page. We should work collaboratively on that wording and submit
it in follow-up patches.