[ATSPI] Load node module locally in `dump_tree_atspi.js`
Currently, we are loading `atspi_inspect.node` from its target location
under `<BUILD>/lib/atspi`. Following what we did for Python, this patch
loads the module from the same directory as the example itself
(<BUILD>/examples/atspi), creating a local symlink of the module.

What this change does in practice is to centralize the knowledge of
the build tree structure in CMake files, rather than having the logic
spread between cmake files and code in the examples; with the hope
of making it more maintainable.