Atspi: Flesh out tree with more information
Include the following in the dump of the native, python, and node tools:
* Description
* States
* Interfaces
* Object Attributes (as opposed to text attributes, from the text iface)
* Relations (see note below)

In addition, there is python and node.js API support to obtain the
above information.

Note: Most objects don't have any relations unless something like
aria-labelledby, aria-describedby, et al. is present on the element.
So to verify the relations functionality in a giant tree dump, try:

  $ examples/atspi/dump_tree_atspi <args> | grep -B4 Relation

Related to issue #138. (This does not provide wrappers for other
interfaces, but the work here is a first step in fleshing out the
tree and providing needed API.)