[ATSPI] Add means to get the target of a given relationship
While this functionality is useful in general, this change provides a
means to quickly get to the document from the frame/window in (at least)
gecko-based and chromium-based browsers. It will be helpful in a future
change to optionally only dump the document tree (i.e. without including
the browser UI), and it provides a means to demonstrate and document the
use of the wrapped interfaces (which in some cases are only implemented
for document content).