Add basic unit-test cmake infra and one (smoke) test example for AT-SPI backend
It is implemented using CTest infrastructure:
https://cmake.org/cmake/help/book/mastering-cmake/chapter/Testing%20With%20CMake%20and%20CTest.html

The idea is for this to be used as boilerplate/placeholder for adding
useful unit-tests in the future. We will need some API to craft
accessibility trees out of thin air (e.g, without depending on a running app)
to then start adding meaningful unit-tests against them.

Runnig the tests is as simple as a `make test`.

The patch includes:

- A new cmake feature flag AXA_UNIT_TESTS is added, ON by default.
- A new top-level `tests` dir is added, following the same structure as
  `examples` (one sub-dir for each backend).
- A single unit test `test_smoke` is added as boilerplate for the test
  infra. As with examples, there are 3 variants: a native, a Python and
  a NodeJS version.
- README updated with a new section `Testing`, and mention of the new feature flag.

Fixes #48 <https://github.com/Igalia/AXAccess/issues/48>